colorout.cprompt: print prompt instead of write prompt
i.e., print a newline at the end of the prompt.

In iTerm2 (didn't test elsewhere), when a prompt is written without
newline, and user enters C-a (Emacs sense) when editing the line, the
cursors returns to the beginning of the prompt (!) rather than the
beginning of user input, and C-e doesn't return to the end, only
forwarding by the length of user input. Internal representation is
correct though. Not sure if this is a bug in Python STL, readline, or
the terminal emulator (iTerm2).

Anyway, start a newline for user input also makes copying user input
easier in iTerm2 (just highlight the whole line), so why not.