config: bug fix
The initializer in the parent class should not set self._conf to None,
since a child class might set it first before calling
super.__init__() (which is the case for INIConfig). Now self._conf is
instead set to None in _get_config_file, which should be implemented by
a subclass anyway.