pbar: support for proprocessed portion
The preprocessed portion counts towards the percentage processed, but
does not count towards speed. Fixes #4.

Also, the progress bar is now drawn at the outset (0%, 0B/s, ETA
unknown) instead after at least the first chunk is processed. This is
because for slow processing, it is good to at least print something to
inform the user that processing has started.